+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.8
+
+
+
 ## 0.0.7
 
 - Make header rows sticky in the Data Editor and Results panels